msm: sps: add option to constrain logging for BAM HW enable event
When a BAM device is enabled, we have log to show the presence of BAM HW.
This logging works fine for most BAMs. However, A2 BAM may be enabled and
disabled frequently when A2 power collapse is enabled. And thus we may have
frequent logging output for A2 BAM. Add the logging option in BAM property
structure so that we could constrain the logging output for A2 BAM.

Change-Id: I14f32f02d32cdf9c6fd4d16f6517ee1305a3ae13
Signed-off-by: Yan He <yanhe@codeaurora.org>